ML_prediction_price/requirements.txt
pandas>=2.0
numpy>=1.24
catboost>=1.2
scikit-learn>=1.3
python-dotenv>=1.0
openai>=1.0
